pyproject.toml
[project]
name = "caresight"
version = "0.1.0"
description = "AI-powered virtual caregiver system for dementia and Alzheimer's patients"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "python-multipart",
    "httpx",
    "motor",
    "google-genai",
    "elevenlabs",
    "websockets",
    "twilio",
    "pyaudio",
    "viam-sdk[mlmodel]",
    "python-dotenv",
    "pydantic-settings",
]
